+++ src/guile/ares/suitbl/core.scm
@@ -13,6 +13,7 @@
                test-loader
                test-thunk
                suite suite?
+               suite-loader suite-loader?
                suite-thunk suite-thunk?
 
                define-suite))